database.py is a “compatibility shim” + legacy CRUD + auth/OTP logic
