+++ db/case_service.py
@@ -1,0 +1,293 @@
+from __future__ import annotations
+
+import datetime as dt
+from typing import List, Optional
+
+from sqlalchemy.orm import Session
+
+from db.models import (
+    Case,
+    CaseDeadline,
+    CaseDocument,
+    CaseOutcome,
+    CaseRecord,
+    CaseStatus,
+    CaseTimeline,
+    ModelFeedback,
+    UserFeedback,
+)
+
+
+def create_case(db: Session, user_id: int, case_number: str, case_type: str, jurisdiction: str, title: Optional[str] = None) -> Case:
+    """Create a new case"""
+    case = Case(
+        user_id=user_id,
+        case_number=case_number,
+        case_type=case_type,
+        jurisdiction=jurisdiction,
+        title=title,
+    )
+    db.add(case)
+    db.commit()
+    db.refresh(case)
+    return case
+
+
+def get_user_cases(db: Session, user_id: int) -> List[Case]:
+    """Get all cases for a user"""
+    return db.query(Case).filter(Case.user_id == user_id).order_by(Case.created_at.desc()).all()
+
+
+def get_case_by_id(db: Session, case_id: int) -> Optional[Case]:
+    """Get a case by ID"""
+    return db.query(Case).filter(Case.id == case_id).first()
+
+
+def get_case_by_number(db: Session, user_id: int, case_number: str) -> Optional[Case]:
+    """Get a case by case number for a specific user"""
+    return db.query(Case).filter(
+        Case.user_id == user_id,
+        Case.case_number == case_number,
+    ).first()
+
+
+def update_case_status(db: Session, case_id: int, status: CaseStatus) -> Optional[Case]:
+    """Update case status"""
+    case = db.query(Case).filter(Case.id == case_id).first()
+    if case:
+        case.status = status
+        db.commit()
+        db.refresh(case)
+    return case
+
+
+def delete_case(db: Session, case_id: int) -> bool:
+    """Delete a case and all related data"""
+    case = db.query(Case).filter(Case.id == case_id).first()
+    if case:
+        db.delete(case)
+        db.commit()
+        return True
+    return False
+
+
+def create_case_document(
+    db: Session,
+    case_id: int,
+    document_type,
+    user_id: int,
+    document_content: Optional[str] = None,
+    file_path: Optional[str] = None,
+    summary: Optional[str] = None,
+    remedies: Optional[dict] = None,
+) -> CaseDocument:
+    """Create a new case document"""
+    doc = CaseDocument(
+        case_id=case_id,
+        document_type=document_type,
+        document_content=document_content,
+        file_path=file_path,
+        summary=summary,
+        remedies=remedies,
+    )
+    db.add(doc)
+    db.commit()
+    db.refresh(doc)
+    return doc
+
+
+def get_case_documents(db: Session, case_id: int) -> List[CaseDocument]:
+    """Get all documents for a case"""
+    return db.query(CaseDocument).filter(
+        CaseDocument.case_id == case_id
+    ).order_by(CaseDocument.uploaded_at).all()
+
+
+def get_case_document_by_id(db: Session, document_id: int) -> Optional[CaseDocument]:
+    """Get a document by ID"""
+    return db.query(CaseDocument).filter(CaseDocument.id == document_id).first()
+
+
+def update_case_document(
+    db: Session,
+    document_id: int,
+    document_content: Optional[str] = None,
+    summary: Optional[str] = None,
+    remedies: Optional[dict] = None,
+) -> Optional[CaseDocument]:
+    """Update case document"""
+    doc = db.query(CaseDocument).filter(CaseDocument.id == document_id).first()
+    if doc:
+        if document_content is not None:
+            doc.document_content = document_content
+        if summary is not None:
+            doc.summary = summary
+        if remedies is not None:
+            doc.remedies = remedies
+        db.commit()
+        db.refresh(doc)
+    return doc
+
+
+def create_case_record(
+    db: Session,
+    hashed_case_id: str,
+    case_type: str,
+    jurisdiction: str,
+    court_name: Optional[str] = None,
+    judge_name: Optional[str] = None,
+    plaintiff_type: Optional[str] = None,
+    defendant_type: Optional[str] = None,
+    case_value: Optional[str] = None,
+    outcome: Optional[str] = None,
+    judgment_summary: Optional[str] = None,
+) -> CaseRecord:
+    """Create a new case record for analytics"""
+    case = CaseRecord(
+        hashed_case_id=hashed_case_id,
+        case_type=case_type,
+        jurisdiction=jurisdiction,
+        court_name=court_name,
+        judge_name=judge_name,
+        plaintiff_type=plaintiff_type,
+        defendant_type=defendant_type,
+        case_value=case_value,
+        outcome=outcome,
+        judgment_summary=judgment_summary,
+    )
+    db.add(case)
+    db.commit()
+    db.refresh(case)
+    return case
+
+
+def get_case_record(db: Session, hashed_case_id: str) -> Optional[CaseRecord]:
+    """Get a case record by hashed ID"""
+    return db.query(CaseRecord).filter(CaseRecord.hashed_case_id == hashed_case_id).first()
+
+
+def get_cases_by_criteria(db: Session, **criteria) -> List[CaseRecord]:
+    """Search case records by criteria"""
+    query = db.query(CaseRecord)
+    for key, value in criteria.items():
+        if hasattr(CaseRecord, key) and value:
+            query = query.filter(getattr(CaseRecord, key) == value)
+    return query.all()
+
+
+def update_case_outcome(
+    db: Session,
+    hashed_case_id: str,
+    appeal_filed: bool = False,
+    appeal_date: Optional[dt.datetime] = None,
+    appeal_outcome: Optional[str] = None,
+    appeal_success: Optional[bool] = None,
+    time_to_appeal_verdict: Optional[int] = None,
+    appeal_cost: Optional[str] = None,
+    additional_notes: Optional[str] = None,
+) -> CaseOutcome:
+    """Update or create case outcome data"""
+    record = get_case_record(db, hashed_case_id)
+    if not record:
+        raise ValueError(f"Case {hashed_case_id} not found")
+
+    outcome = db.query(CaseOutcome).filter(CaseOutcome.case_id == record.id).first()
+    if not outcome:
+        outcome = CaseOutcome(case_id=record.id)
+        db.add(outcome)
+
+    outcome.appeal_filed = appeal_filed
+    if appeal_date:
+        outcome.appeal_date = appeal_date
+    if appeal_outcome:
+        outcome.appeal_outcome = appeal_outcome
+    if appeal_success is not None:
+        outcome.appeal_success = appeal_success
+    if time_to_appeal_verdict:
+        outcome.time_to_appeal_verdict = time_to_appeal_verdict
+    if appeal_cost:
+        outcome.appeal_cost = appeal_cost
+    if additional_notes:
+        outcome.additional_notes = additional_notes
+
+    db.commit()
+    db.refresh(outcome)
+    return outcome
+
+
+def get_user_stats(db: Session, user_id: int) -> dict:
+    """Calculate high-level stats for a user dashboard"""
+    cases = get_user_cases(db, user_id)
+
+    active_count = len([c for c in cases if c.status == CaseStatus.ACTIVE])
+    appealed_count = len([c for c in cases if c.status == CaseStatus.APPEALED])
+    closed_count = len([c for c in cases if c.status == CaseStatus.CLOSED])
+
+    now = dt.datetime.now(dt.timezone.utc)
+    upcoming_deadlines = db.query(CaseDeadline).filter(
+        CaseDeadline.user_id == user_id,
+        CaseDeadline.is_completed == False,
+        CaseDeadline.deadline_date > now,
+    ).count()
+
+    return {
+        "total_cases": len(cases),
+        "active_cases": active_count,
+        "appealed_cases": appealed_count,
+        "closed_cases": closed_count,
+        "upcoming_deadlines": upcoming_deadlines,
+    }
+
+
+def submit_model_feedback(
+    db: Session,
+    user_id: str,
+    model_name: str,
+    task: str,
+    case_id: Optional[int] = None,
+    is_accurate: Optional[bool] = None,
+    corrected_text: Optional[str] = None,
+    feedback_notes: Optional[str] = None,
+) -> ModelFeedback:
+    """Submit model output feedback"""
+    fb = ModelFeedback(
+        user_id=str(user_id),
+        model_name=model_name,
+        task=task,
+        case_id=case_id,
+        is_accurate=is_accurate,
+        corrected_text=corrected_text,
+        feedback_notes=feedback_notes,
+    )
+    db.add(fb)
+    db.commit()
+    db.refresh(fb)
+    return fb
+
+
+def get_case_timeline(db: Session, case_id: int) -> List[CaseTimeline]:
+    """Get all timeline events for a case"""
+    return db.query(CaseTimeline).filter(CaseTimeline.case_id == case_id).order_by(CaseTimeline.event_date.desc()).all()
+
+
+def create_timeline_event(
+    db: Session,
+    case_id: int,
+    event_type: str,
+    description: str,
+    event_date: Optional[dt.datetime] = None,
+    metadata: Optional[dict] = None,
+) -> CaseTimeline:
+    """Create a new timeline event"""
+    event = CaseTimeline(
+        case_id=case_id,
+        event_type=event_type,
+        description=description,
+        event_date=event_date or dt.datetime.now(dt.timezone.utc),
+        event_metadata=metadata,
+    )
+    db.add(event)
+    db.commit()
+    db.refresh(event)
+    return event
+
