+++ db/notifications_service.py
@@ -1,0 +1,67 @@
+from __future__ import annotations
+
+import datetime as dt
+from typing import Optional
+
+from sqlalchemy.orm import Session
+
+from db.models import CaseDeadline, NotificationChannel, NotificationTemplate, UserPreference
+
+
+def create_or_update_user_preference(
+    db: Session,
+    user_id: int,
+    email: str,
+    phone_number: Optional[str] = None,
+    notification_channel: NotificationChannel = NotificationChannel.BOTH,
+    timezone: str = "UTC",
+    holiday_aware_reminders: bool = False,
+    holiday_country: Optional[str] = None,
+    holiday_region: Optional[str] = None,
+    holiday_calendar_json: Optional[str] = None,
+) -> UserPreference:
+    """Create or update user notification preferences"""
+    pref = db.query(UserPreference).filter(UserPreference.user_id == user_id).first()
+
+    if pref:
+        pref.email = email
+        pref.phone_number = phone_number
+        pref.notification_channel = notification_channel
+        pref.timezone = timezone
+        pref.holiday_aware_reminders = holiday_aware_reminders
+        pref.holiday_country = holiday_country
+        pref.holiday_region = holiday_region
+        pref.holiday_calendar_json = holiday_calendar_json
+        pref.updated_at = dt.datetime.now(dt.timezone.utc)
+    else:
+        pref = UserPreference(
+            user_id=user_id,
+            email=email,
+            phone_number=phone_number,
+            notification_channel=notification_channel,
+            timezone=timezone,
+            holiday_aware_reminders=holiday_aware_reminders,
+            holiday_country=holiday_country,
+            holiday_region=holiday_region,
+            holiday_calendar_json=holiday_calendar_json,
+        )
+        db.add(pref)
+
+    db.commit()
+    db.refresh(pref)
+    return pref
+
+
+def get_notification_template_for_user(db: Session, user_id: int) -> Optional[NotificationTemplate]:
+    """Get notification template for a user"""
+    return db.query(NotificationTemplate).filter(NotificationTemplate.user_id == user_id).first()
+
+
+def get_user_deadlines(db: Session, user_id: int):
+    """Get all active deadlines for a user"""
+    now = dt.datetime.now(dt.timezone.utc)
+    return db.query(CaseDeadline).filter(
+        CaseDeadline.user_id == user_id,
+        CaseDeadline.is_completed == False,
+        CaseDeadline.deadline_date > now,
+    ).order_by(CaseDeadline.deadline_date).all()
